requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.2.1
antlr4-python3-runtime==4.9.3
anyio==3.6.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
-e git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1#egg=clip
coloredlogs==15.0.1
diffusers==0.2.4
einops==0.4.1
fastapi==0.82.0
filelock==3.8.0
flatbuffers==2.0.7
frozenlist==1.3.1
fsspec==2022.8.2
ftfy==6.1.1
google-auth==2.11.0
google-auth-oauthlib==0.4.6
greenlet==1.1.3
grpcio==1.48.1
h11==0.13.0
huggingface-hub==0.9.1
humanfriendly==10.0
idna==3.3
imageio==2.21.2
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
invisible-watermark==0.1.5
jedi==0.18.1
joblib==1.1.0
kafka-python==2.0.2
kornia==0.6.7
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
mpmath==1.2.1
multidict==6.0.2
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.0
omegaconf==2.2.3
onnx==1.12.0
onnxruntime==1.12.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
parso==0.8.3
Pillow==9.2.0
protobuf==3.19.4
psycopg2==2.7.7
pudb==2022.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pyDeprecate==0.3.2
Pygments==2.13.0
pyparsing==3.0.9
pytorch-lightning==1.7.4
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
sympy==1.10.1
-e git+https://github.com/CompVis/taming-transformers.git@24268930bf1dce879235a7fddd0b2355b84d7ea6#egg=taming_transformers
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.12.1
torch==1.12.1
torch-fidelity==0.3.0
torchaudio==0.8.2
torchmetrics==0.9.3
torchvision==0.13.1
tqdm==4.64.1
transformers==4.19.2
typing_extensions==4.3.0
urllib3==1.26.12
urwid==2.1.2
urwid-readline==0.13
uvicorn==0.18.3
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
